bai3

+++ bai3/src/main/java/com/example/bai2/model/Book.java
@@ -2,24 +2,23 @@ package com.example.bai2.model;
 
 public class Book {
 
-    private int id;
+    private Long id;
     private String title;
     private String author;
 
-    public Book() {
-    }
+    public Book() {}
 
-    public Book(int id, String title, String author) {
+    public Book(Long id, String title, String author) {
         this.id = id;
         this.title = title;
         this.author = author;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
